extract getOptions to module
